requirements.txt
Django==1.11.6
django-cors-headers==2.1.0
django-filter==1.1.0
django-private-storage==1.2.2
django-rest-swagger==2.1.2
djangorestframework-jwt==1.11.0
drf-openapi==1.0.0
Pillow==4.3.0
drf_generators